List, delete files from S3 + sample registration validation

+++ src/api/services/DataSubmissionService.ts
@@ -6,9 +6,6 @@ import { Logger, LoggerInterface } from '../../decorators/Logger';
 import { DataSubmissionRepository } from '../repositories/DataSubmissionRepository';
 import { DataSubmission } from '../models/DataSubmission';
 import { SampleRegistrationRepository } from '../repositories/SampleRegistrationRepository';
-import { SampleRegistration } from '../models/SampleRegistration';
-import { TypedDataRecord } from '@overturebio-stack/lectern-client/lib/schema-entities';
-import { Status, User } from '../models/ReferentialData';
 import { events } from '../subscribers/events';
 
 @Service()
@@ -55,6 +52,11 @@ export class DataSubmissionService {
     public async update(dataSubmission: DataSubmission): Promise<DataSubmission> {
         this.log.debug('Updating a data submission');
 
+        const initialDataSubmission: DataSubmission = await this.findOne(dataSubmission.id);
+        if (!initialDataSubmission) {
+            throw new Error(`Data submission with ${dataSubmission.id} does not exist.`);
+        }
+
         // Delete all registered samples and recreate.
         await this.sampleRegistrationRepository.delete({
             dataSubmissionId: dataSubmission.id,
@@ -73,41 +75,14 @@ export class DataSubmissionService {
         return this.dataSubmissionRepository.save(dataSubmission);
     }
 
-    public async createOrUpdate(
-        dataSubmissionId: number = NaN,
-        validatedRecords: TypedDataRecord[],
-        user: User
-    ): Promise<DataSubmission> {
-        let dataSubmission: DataSubmission;
-
-        const registeredSamples: SampleRegistration[] = validatedRecords.map((row) => new SampleRegistration(row));
-
-        if (!dataSubmissionId || isNaN(dataSubmissionId)) {
-            dataSubmission = new DataSubmission();
-            dataSubmission.status = Status.IN_PROGRESS;
-
-            dataSubmission.createdBy = user.id;
-            dataSubmission.registeredSamples = registeredSamples;
-            try {
-                dataSubmission = await this.create(dataSubmission);
-            } catch (err) {
-                return new Promise(undefined);
-            }
-        } else {
-            dataSubmission = await this.findOne(dataSubmissionId);
-            if (!dataSubmission) {
-                throw new Error(`Data submission with ${dataSubmissionId} does not exist.`);
-            }
-            dataSubmission.lastUpdatedBy = user.id;
-            dataSubmission.registeredSamples = registeredSamples;
-            dataSubmission = await this.update(dataSubmission);
-        }
-
-        return dataSubmission;
-    }
-
     public async delete(id: number): Promise<void> {
         this.log.info('Delete a data submission');
+
+        // Clean up all registered samples
+        await this.sampleRegistrationRepository.delete({
+            dataSubmissionId: id,
+        });
+
         await this.dataSubmissionRepository.delete(id);
         return;
     }
